+++ python/cdp-langchain/cdp_langchain/agent_toolkits/cdp_toolkit.py
@@ -67,6 +67,8 @@ class CdpToolkit(BaseToolkit):
             wow_buy_token
             wow_sell_token
             wrap_eth
+            get_price_inference
+            get_all_topics
     Use within an agent:
         .. code-block:: python